Move share interfaces to own files so they can be autoloaded
